+++ src/plugin.cpp
@@ -17,6 +17,7 @@ void init(Plugin *p) {
 	p->addModel(modelAuxSends);
 	p->addModel(modelBitcrusher);
 	p->addModel(modelDrumBus);
+	p->addModel(modelTapeDelay);
 
 	// Any other plugin initialization may go here.
 	// As an alternative, consider lazy-loading assets and lookup tables when your module is created to reduce startup times of Rack.