+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## v2.0.3
 
+- Add TapeDelay module
 - Fix small memory leak in BitCrusher and DJFilter
 
 ## v2.0.2